Rename messaging-next feature to conversations
The '-next' suffix outlived the older version it distinguished from, and
'messaging' was a second name for a concept the rest of the codebase
(and backend: ConversationChannel DO, ConversationRow, conversations-client,
selectedConversationStore) already calls 'conversation'.

- src/features/messaging-next/ -> src/features/conversations/
- Update import sites (MainContent, stores, main.css @import)
- Flip the MainContent ViewMode token 'messaging' -> 'conversations'
  (visible nav label stays t('nav.messages'); no CSS/i18n coupling)
- localStorage draft key hangvidu:messaging-next:draft -> hangvidu:conversations:draft
- Refresh in-tree comments, test labels, STATE_RULES doc row

playgrounds/boundary-imports.json left stale (regenerable scan artifact).
'message' (the item inside a conversation) is a distinct concept and unchanged.
