+++ src/styles/main.css
@@ -23,7 +23,7 @@
 @import './components/videos.css';
 @import './components/chat-controls.css';
 @import './components/messages.css';  */
-@import '../features/messaging-next/reactions/reactions.css';
+@import '../features/conversations/reactions/reactions.css';
 @import './components/search.css';
 @import './components/notifications.css';
 @import './components/toast.css';
